Added a new view to show the URI and SQL being used
Added a custom view that you can set the sql and the uri on. The uri is set by a listener for when the
uri is being created it will be passed to the view. I'm thinking possibly could be a context leak by
passing the view around but figuring this is a demo, will get away with it. The SQL at the moment is set
manually, no way to query the SQLiteProvider for the SQL at the moment