Updating the README for updating from old versions
### Problem

We were on 1.0.4 and not seeing any newer version because the name changed from sqliteprovider-core to sqliteprovider. Then when I came here and saw the latest version was 1.2.0 I just tried changing the number and was wondering why it didn't work.

### Solution

Just adding a little note to the README to draw attention to that name change, in the case that anyone is upgrading from a super old version (1.0.4).